Add analyzer and codefix to improve setup scopes
We added the parameterless `mock.Setup()` to make it more discoverable how to switch the entire mock(s) to setup mode so the invocation tracking and recording and strict mock validation doesn't kick in, while you still get to use the lamba-free way of setting things up.

That pattern, however, could be confusing and lead you to believe you need one such Setup() call for every mock being set up in the using block. This is not the case: the setup scope is active for the entire call-context/thread, so there's no need for the method to be called on a particular instance of one mock.

This analyzer and codefix makes this more obvious and also introduces users to the Syntax static helper, which should make the test code also much more succint since it has optimized members for shorter syntax (i.e. for Verify, Raise, Any/Not and so on).

The added helpers are mostly brought as-is from Stunts, but with a simplified folder structure for individual analyzer codes.