Add codefix for setup of ref/out methods
This is one of the most annoying scenarios when mocking APIs that have ref/out arguments. Figuring out the callback parameters, setting up the right output values in an untyped collection of some sort, it's terrible.

So rather than that, offer a built-in way to generate a delegate that can be used to directly set up and implement the method, with full typed arguments including ref/out annotations, just like you would if you implemented the interface manually (which you likely do when things get complicated because of this issue!).

For easier troubleshooting a VSIX project is also provided so you can more easily debug the completion provider.

Setting up tests for this will be more complicated than for built-in analyzers, pending further investigation.