Add support for both ReturnsAsync and Throws for async methods
ReturnsAsync makes sense since you'd typically want to pass in the actual value and have Moq wrap it in a Task<T>. For Throws, however, since you're not providing any values to it, I don't think it makes sense to force the distinction between Throws and ThrowsAsync. After all, both would be void-returning methods anyway, and you wouldn't want to be forced to await the result either (like an Assert.ThrowsAsync in xunit, say). So for Throws we added overloads for Task/ValueTask and made the implementation smarter so that it can properly detect when a return type of Task<T>/ValueTask<T> is expected and wrap the exception appropriately.